+++ layers/GatoGraphQLForWP/plugins/gatographql/gatographql.php
@@ -57,6 +57,30 @@ if (class_exists(Plugin::class) && !PluginApp::getMainPluginManager()->assertIsV
 }
 
 /**
+ * Validate that there is enough memory to run the plugin.
+ *
+ * Minimum: 64MB
+ */
+$minRequiredPHPMemoryLimit = '64M';
+$minRequiredPHPMemoryLimitInBytes = \wp_convert_hr_to_bytes($minRequiredPHPMemoryLimit);
+$phpMemoryLimit = \ini_get('memory_limit');
+$phpMemoryLimitInBytes = \wp_convert_hr_to_bytes($phpMemoryLimit);
+if ($phpMemoryLimitInBytes < $minRequiredPHPMemoryLimitInBytes) {
+    \add_action('admin_notices', function () use ($minRequiredPHPMemoryLimit, $phpMemoryLimit) {
+        printf(
+            '<div class="notice notice-error"><p>%s</p></div>',
+            sprintf(
+                __('Plugin <strong>%1$s</strong> requires at least <strong>%2$s</strong> of memory, however the server\'s PHP memory limit is set to <strong>%3$s</strong>. Please increase the memory limit to load %1$s.', 'gatographql'),
+                __('Gato GraphQL', 'gatographql'),
+                $minRequiredPHPMemoryLimit,
+                $phpMemoryLimit
+            )
+        );
+    });
+    return;
+}
+
+/**
  * The commit hash is added to the plugin version 
  * through the CI when merging the PR.
  *